requirements.txt
blessings==1.6
click==6.7
czech-holidays==0.1.3
gitdb2==2.0.0
GitPython==2.1.1
oauthlib==2.0.1
python-dateutil==2.6.0
pyvodb==0.3.3
PyYAML==3.12
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
smmap2==2.0.1
SQLAlchemy==1.1.4
tweepy==3.5.0
Unidecode==0.4.20
